--- src/main/java/woowacourse/auth/ui/AuthControllerAdvice.java
@@ -1,33 +1,0 @@
-package woowacourse.auth.ui;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import woowacourse.auth.dto.ErrorResponse;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
-import woowacourse.shoppingcart.exception.InvalidTokenException;
-
-@RestControllerAdvice
-public class AuthControllerAdvice {
-
-    @ExceptionHandler(InvalidCustomerException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidCustomerException(Exception e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ErrorResponse> handleIllegalArgumentException(Exception e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
-
-    @ExceptionHandler(InvalidTokenException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidTokenException(Exception e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(e.getMessage()));
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Void> handle() {
-        return ResponseEntity.badRequest().build();
-    }
-}
